Pass the sanitized args to thin client in sbt.bat
Passing the unsanitized arguments was problematic in sbt.bat because it
caused sbt --client to run sbtn with --client as a parameter. This
caused sbtn to launch sbt with the --client which caused a loop. This
manifested as a weird error about not being able to get the console
mode.