Use last modified for stamping resource files
We don't necessarily need to hash resource files. The only advantage to
hashing them is that a continuous build won't trigger if the content of
the resource file doesn't change. The drawback to hashing them is that
resource files can be much bigger than a typical source file and in the
worst case can cause an oom because FarmHash loads the entire file into
memory to perform the hashing.