Add scripted option to toggle sbt instance reuse
In sbt 1.3.x, every scripted test got its own server. I am not exactly
sure why that behavior changed in 1.4.0, but we now by default reuse the
same sbt server in scripted tests. This is a very nice performance
benefit that I think most builds benefit from. However, there are some
cases where reusing the sbt server can cause problems. In windows, in
particular, it may be the case that there are files that sbt is not able
to delete in the project directory that may leak from one scripted test
into another. To avoid this, this commit adds the scriptedReuseServer
parameter which when set to false allows the user to rever their build
to the sbt 1.3.x semantics which are much slower.

This came from #6042. With reuseServer true, the paradox scripted tests
take about 1 minute on my computer. With reuseServer false, they take
about 5 minutes.