Use NetworkClient to implement `sbt -bsp`
Network client already supports the -bsp command (since
65ab7c94d038955d48eefbf7951fb9cd399ab7ef). This commit reworks the
BspClient.run method so that it delegates to the NetworkClient. The
advantage to doing it this way is that improvements to starting up the
sbt server by the thin client will automatically propagate to the -bsp
command. The way that it is implemented, all of the output generated
during server startup will be redirected to System.err which is useful
for debugging without messing up the bsp protocol, which relies on only
bsp messages being written to System.out.