Optionally provide sbt jline terminal for some tasks
The dotty repl task doesn't work with sbt 1.4.x when started with
in-process runMain. There were two problems:
1. the top classloader for run was not the jline 3 classloader so sbt
   was unable to override TerminalBuilder.build
2. even when setting the top loader to the jline 3 classloader, we
   weren't actually setting the global value so that it could be
   consumed by TerminalBuilder.build

I didn't want to create a JLine3 terminal for every evaluation of
MainLoop.process so I added a setting, needsJLine3, which provides a
list of tasks for which we need to setup a jline 3 terminal before
evaluation.

In addition to these changes, the dotty build.sbt needs to add some
settings:

inProcessTopClassLoader := scalaInstanceTopLoader.value,
Global / needsJLine3 += "repl",
and in the repl task, it needs to add
terminal.value.setMode(canonical = false, echo = false)