Fixes M1 support
Fixes https://github.com/sbt/sbt/issues/7117

Problem
-------
sbt no longer works on a Mac with M1 (aarch64) chips because it fails on
ipcsocket library, which we maintain.
This was caused by architecture detection that was added in 1.6.1,
which did not take in account for macOS universal binary.

Solution
--------
Update to ipcsocket 1.6.2.